Random changes and bugfixes.
 * Expires headers added for assets.
 * Corrected adding to the watches table and the unwatching can now
   receive a user ID. This means watches can be added and removed
   correctly.
 * Corrected naming of base_url. Stupid mistake.
 * The code for rendering watches hasn't needed to check for a null
   assigned_user_id for ages, so I trimmed that out of the code.
 * Added an extra priority for documentation.
 * Until (and after) I actually add title validation to the forms, the
   issue ID can be clicked on.

+++ classes/Issues.php
@@ -82,15 +82,18 @@ class Issues {
 		}
 
 		// Just in case...
-		self::unwatch($id);
+		self::unwatch($issue_id, $user_id);
 
 		$db->insert('watches', compact('user_id', 'issue_id'));
 	}
 
-	public static function unwatch($id) {
+	public static function unwatch($issue_id, $user_id=null) {
 		global $db;
-		$db->execute("DELETE FROM watches WHERE user_id = %d AND issue_id = %d",
-			Users::current()->get_id(), $id);
+		if (is_null($user_id)) {
+			$user_id = Users::current()->get_id();
+		}
+
+		$db->execute("DELETE FROM watches WHERE user_id = %d AND issue_id = %d", $user_id, $issue_id);
 	}
 
 	public static function get_watchers($id) {
